🔧 Rename yearbook components folder to characters
