+++ app/components/characters/ProfileCard/index.tsx
