+++ app/components/characters/ProfileCard/styles/LowerNotch.tsx
