+++ app/components/characters/ProfileCard/styles/ProfilePicture.tsx
