+++ app/components/characters/ProfileCard/styles/UpperNotch.tsx
