+++ app/components/characters/ProfileCard/styles/index.tsx
