+++ app/components/characters/SearchBar/SearchBarButton.tsx
