+++ app/components/characters/SearchBar/SearchBarDropdown.tsx
