+++ app/components/characters/SearchBar/index.tsx
