+++ app/components/characters/SearchBar/styles/SearchButton.tsx
