+++ app/components/characters/SearchBar/styles/SearchDropdown.tsx
