+++ app/components/characters/SearchBar/styles/SearchInput.tsx
