+++ app/components/characters/SearchBar/styles/index.tsx
