+++ app/components/characters/house-icons/Animation.tsx
