+++ app/components/characters/house-icons/Comedy.tsx
