+++ app/components/characters/house-icons/Fantasy.tsx
