+++ app/components/characters/house-icons/Horror.tsx
