+++ app/components/characters/house-icons/Mystery.tsx
