+++ app/components/characters/house-icons/Space.tsx
