+++ app/components/characters/ts/interfaces/Student.ts
