+++ app/components/characters/ts/interfaces/index.ts
