+++ app/components/characters/utils/helpers.tsx
