+++ app/components/characters/utils/students.ts
